move error related functions to new file "error"
