+++ src/error.js
@@ -1,0 +1,10 @@
+function showError() {
+  $('#error').removeClass('hidden');
+}
+export function handleError(message) {
+  $('#error').html(message);
+  showError();
+}
+export function hideError() {
+  $('#error').addClass('hidden');
+}
